Updated pid cpp and hpp
Reverted previousOutput to a long double. previousOutput is no longer in mV (now in volts).

+++ include/util/pid.hpp
@@ -60,7 +60,7 @@ private:
 
     bool usingProfile {false};
 
-    unsigned int previousOutput {0};
+    long double previousOutput {0};
 
 };
